+++ CHANGES.rst
@@ -13,6 +13,7 @@ Unreleased
 - :meth:`.Redditor.moderated` will now objectify all data returned from the API.
 - The `wiki_edit` endpoint has been changed from `r/{subreddit}/api/wiki/edit/` to
   `r/{subreddit}/api/wiki/edit`.
+- :meth:`.Redditor.block` no longer needs to retrieve a user's fullname.
 
 **Fixed**